music_service/requirements.txt
Django==4.1.7
djangorestframework==3.14.0
drf-yasg==1.21.5
gunicorn==20.0.4
psycopg2-binary==2.8.6